Ensured all MAME processes are terminated on application shutdown, tidied shutdown code if shutting down prior to message loop construction, shifted MAME sanity check before game initialisation.
