+++ Mamesaver/MameOrchestrator.cs
@@ -70,13 +70,6 @@ namespace Mamesaver
                     return;
                 }
 
-                // Start listening for user input events
-                _screenManager.Initialise(_cancellationTokenSource);
-                _hotKeyManager.Initialise();            
-
-                // Start power management timer
-                _powerManager.Initialise();
-
                 // Verify that MAME can be run so we can return immediately if there are errors
                 try
                 {
@@ -92,6 +85,14 @@ namespace Mamesaver
                 // Find the best primary screen for MAME. As games are largely vertical and screens are wide, select the one with the greatest Y axis
                 var bestPrimaryScreen = Screen.AllScreens.OrderByDescending(screen => screen.Bounds.Height).First();
 
+                _screenManager.Initialise(_cancellationTokenSource);
+
+                // Start listening for user input events
+                _hotKeyManager.Initialise();
+
+                // Start power management timer
+                _powerManager.Initialise();
+ 
                 // Initialise primary MAME screen
                 _gamePlayManager.Initialise(bestPrimaryScreen, _cancellationTokenSource);
                 _mameScreen.Initialise(bestPrimaryScreen);
@@ -141,7 +142,9 @@ namespace Mamesaver
             try
             {
                 Log.Information("Stopping screensaver");
-                Application.Exit();
+
+                if (Application.MessageLoop) Application.Exit();
+                else Environment.Exit(1);
             }
             catch (Exception e)
             {
